+++ src/components/MiniPlayer.tsx
@@ -1,5 +1,5 @@
 import { usePlaybackCoverArt } from '../hooks/usePlaybackCoverArt';
-import { resolveSubsonicSongCoverArtId } from '../utils/library/advancedSearchLocal';
+import { resolvePlaybackTrackCoverArtId } from '../cover/resolveCoverArtId';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import { emit } from '@tauri-apps/api/event';
 import { invoke } from '@tauri-apps/api/core';
@@ -141,17 +141,7 @@ export default function MiniPlayer() {
   }, [queueOpen, state.queueIndex]);
 
   const { track, isPlaying } = state;
-  const miniCoverArtId = useMemo(
-    () =>
-      track
-        ? resolveSubsonicSongCoverArtId({
-            id: track.id,
-            coverArt: track.coverArt,
-            albumId: track.albumId ?? '',
-          })
-        : undefined,
-    [track],
-  );
+  const miniCoverArtId = useMemo(() => resolvePlaybackTrackCoverArtId(track), [track]);
   const { src: miniCoverSrc, cacheKey: miniCoverKey } = usePlaybackCoverArt(miniCoverArtId, 300);
   const progress = duration > 0 ? Math.min(100, (currentTime / duration) * 100) : 0;
