+++ src/components/artistDetail/ArtistSuggestionTrackCover.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { CoverArtImage } from '../../cover/CoverArtImage';
-import { COVER_DENSE_ARTIST_LIST_CSS_PX } from '../../cover/layoutSizes';
+import { COVER_ARTIST_TOP_TRACK_CSS_PX } from '../../cover/layoutSizes';
 
 export default function ArtistSuggestionTrackCover({ coverArt, album }: { coverArt: string; album: string }) {
   return (
     <CoverArtImage
       coverArtId={coverArt}
-      displayCssPx={COVER_DENSE_ARTIST_LIST_CSS_PX}
+      displayCssPx={COVER_ARTIST_TOP_TRACK_CSS_PX}
       surface="sparse"
       ensurePriority="high"
       alt={album}
