+++ src/utils/library/advancedSearchLocal.ts
@@ -154,37 +154,6 @@ export function resolveTrackCoverArtId(
   return undefined;
 }
 
-/**
- * Subsonic songs often set `coverArt` to the track id (no art). Prefer `albumId` then.
- */
-export function resolveSubsonicSongCoverArtId(
-  song: Pick<SubsonicSong, 'id' | 'coverArt' | 'albumId'>,
-): string | undefined {
-  const albumId = song.albumId?.trim();
-  const cover = song.coverArt?.trim();
-  if (cover && song.id && cover === song.id && albumId) return albumId;
-  if (albumId) return albumId;
-  if (cover) return cover;
-  return undefined;
-}
-
-/**
- * Artist top tracks: use the album row's `coverArt` when the grid already warmed it.
- */
-export function resolveArtistPageSongCoverArtId(
-  song: Pick<SubsonicSong, 'id' | 'coverArt' | 'albumId' | 'album'>,
-  albums: ReadonlyArray<Pick<SubsonicAlbum, 'id' | 'name' | 'coverArt'>>,
-): string | undefined {
-  const album = song.albumId
-    ? albums.find(a => a.id === song.albumId)
-    : albums.find(a => a.name === song.album);
-  const albumCover = album?.coverArt?.trim();
-  const songId = song.id?.trim();
-  // Album row `coverArt` can echo the track id (no art) — do not prefer it over albumId.
-  if (albumCover && (!songId || albumCover !== songId)) return albumCover;
-  return resolveSubsonicSongCoverArtId(song);
-}
-
 export function trackToSong(t: LibraryTrackDto): SubsonicSong {
   const raw = isObject(t.rawJson) ? t.rawJson : {};
   const resolvedBpm = t.bpm != null && t.bpm > 0 ? t.bpm : undefined;
